+++ CHANGELOG.md
@@ -2,9 +2,21 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [43.0.1] 2026-06-30
+## [43.2.1] 2026-07-07
 ### Fixed
 - Use antd radio buttons in matrix layout.
+
+## [43.2.0] 2026-07-07
+### Added
+- New onFormChange event, emitted when the item value is modified at any
+  level of the form, or when a repeating item is added or removed.
+
+## [43.1.0] 2026-06-30
+### Changed
+- For a multi-select dropdown layout question with sub items, the sub items
+  will be rendered for each selected dropdown option.
+- When exported to QuestionnaireResponse, each sub item will go under its
+  associated answer, in the format of item.answer.item.
 
 ## [43.0.0] 2026-06-29
 ### Changed